Step Three: Customizing the Principal
Customizing the user principal with JWT is similar to what you did in
Module 1 when you customized it with HTTP Basic. Here, you introduce a
bridge object and then return an authentication that supports a custom
principal, which in this case is BearerTokenAuthentication.

Also observe the important difference between user authorities and
client authorities. In the custom JwtAuthenticationConverter, the
authorities are merged since we don't want a client to do something
that the user can't.